Moved the User model into models.py
